Fix first 5 directories of vyper external tests

+++ vyper/external/abiEncoderV2/struct/struct_short.vy
@@ -1,5 +1,5 @@
 struct S:
-    a: int128
+    a: int256
     b: uint256
     c: bytes32
 @external
@@ -10,5 +10,5 @@ def f(s: S) -> S:
 # compileToEwasm: also
 # compileViaYul: also
 # ----
-# f((int256,uint256,bytes16)): 0xff010, 0xff0002, "abcd" -> 0xff010, 0xff0002, "abcd"
-# f((int256,uint256,bytes16)): 0xff010, 0xff0002, 0x1111222233334444555566667777888800000000000000000000000000000000 -> 0xff010, 0xff0002, left(0x11112222333344445555666677778888)
+# f((int256,uint256,bytes32)): 0xff010, 0xff0002, "abcd" -> 0xff010, 0xff0002, "abcd"
+# f((int256,uint256,bytes32)): 0xff010, 0xff0002, 0x1111222233334444555566667777888800000000000000000000000000000000 -> 0xff010, 0xff0002, left(0x11112222333344445555666677778888)
